Updated to using DFG abstraction
