+++ src/SubGraphFunctions.jl
@@ -49,7 +49,7 @@ variable labels `syms::Vector{Symbols}`.
 function buildSubgraphFromLabels(dfg::G, syms::Vector{Symbol}) where G <: AbstractDFG
   cliqSubFg = initfg()
   for sym in syms
-    DFG.GraphsJl.getSubgraphAroundNode(dfg, DFG.GraphsJl.getVariable(dfg, sym), 2, false, cliqSubFg)
+    DFG.getSubgraphAroundNode(dfg, DFG.getVariable(dfg, sym), 2, false, cliqSubFg)
   end
   return cliqSubFg
 end
@@ -57,7 +57,7 @@ end
 #   fgseg = initfg() #sessionname=dfg.sessionname, robotname=dfg.robotname)
 #
 #   for sym in syms
-#     vert = DFG.GraphsJl.getVariable(dfg, sym) #, api=localapi)
+#     vert = DFG.getVariable(dfg, sym) #, api=localapi)
 #     st = getSofttype(vert)
 #     addVariable!(fgseg, sym, st) #, uid=vert.index)
 #     if isInitialized(dfg,sym)
@@ -66,15 +66,15 @@ end
 #   end
 #
 #   for sym in syms
-#     for fct in DFG.GraphsJl.ls(dfg, :x1)
+#     for fct in DFG.ls(dfg, :x1)
 #       if !hasFactor(fgseg, fct)
 #         # check all variables are in desired variable set
-#         possibleVars = DFG.GraphsJl.lsf(dfg, fct)
+#         possibleVars = DFG.lsf(dfg, fct)
 #         ivars = intersect(possibleVars, syms)
 #         @show length(ivars), length(possibleVars)
 #         if length(ivars) == length(possibleVars)
 #           # fvert = getVert(dfg, fct, api=localapi, nt=:fct)
-#           ufc = DFG.GraphsJl.getFactor(dfg, fct) # fvert
+#           ufc = DFG.getFactor(dfg, fct) # fvert
 #
 #           addFactor!(fgseg, possibleVars, getData(ufc).fnc.usrfnc!, autoinit=false) #, uid=fvert.index)
 #         end
