+++ test/unittests/FactorGraph01_tests.jl
@@ -4,8 +4,6 @@ using Juno
 using IncrementalInference
 using RoME
 using DistributedFactorGraphs
-import DistributedFactorGraphs.GraphsJl
-const DFGGraphs = DistributedFactorGraphs.GraphsJl
 using Test
 
 dfg = initfg()
@@ -30,24 +28,32 @@ addFactor!(dfg, [:x2, :l1], deepcopy(p2br), autoinit=false)
 addFactor!(dfg, [:x3, :l2], deepcopy(p2br), autoinit=false)
 
 # Show it
-DFGGraphs.toDotFile(dfg, "/tmp/testRmMarg.dot")
+DFG.toDotFile(dfg, "/tmp/testRmMarg.dot")
 
 # Test with a copy
 dfgPrime = deepcopy(dfg)
-elimOrder = IncrementalInference.getEliminationOrder(dfgPrime)
+elimOrder = getEliminationOrder(dfgPrime)
+# Forcing elimination order to match the documentation
 elimOrder = [:l1, :l2, :x1, :x2, :x3]
 
-IncrementalInference.buildBayesNet!(dfgPrime, elimOrder)
+# tree for elim order  [:l1, :l2, :x1, :x2, :x3]
+# :l1 <-- :x1, :x2
+# :x1 <-- :x2
+# :x2 <-- :x3
+# :l2 <-- :x3
+# :x3 <-- /-/
+
+buildBayesNet!(dfgPrime, elimOrder)
 # Reproduce example
 # IncrementalInference.rmVarFromMarg(dfgPrime, lm[1], )
-DFGGraphs.toDotFile(dfgPrime, "/tmp/testRmMarg.dot")
+DFG.toDotFile(dfgPrime, "/tmp/testRmMarg.dot")
 # Assert that everything was eliminated and every variable has a BayesNetVertID
 @test all(map(v -> getData(v.dfgNode).eliminated, values(dfgPrime.g.vertices)))
-@test all(map(v -> getData(v).BayesNetVertID != nothing, DFGGraphs.getVariables(dfgPrime)))
+@test all(map(v -> getData(v).BayesNetVertID != nothing, DFG.getVariables(dfgPrime)))
 # Assert that we have the expected Bayes tree
 expectedBayesOutVertDict = Dict{Symbol, Vector{Symbol}}(:x2 => [:x3], :l1 => [:x1, :x2], :x3 => [], :l2 => [:x3], :x1 => [:x2])
 for (vId,linkIds) in expectedBayesOutVertDict
-    v = DFGGraphs.getVariable(dfgPrime, vId)
+    v = DFG.getVariable(dfgPrime, vId)
     @test setdiff(getData(v).BayesNetOutVertIDs, linkIds) == []
 end
